feat(): inicio pantalla registrar pago

+++ front/src/services/axiosClient.ts
@@ -33,24 +33,26 @@ axiosClient.interceptors.request.use(
 
 // Función para configurar interceptores con el router
 export const setupAxiosInterceptors = (router: any) => {
-  axiosClient.interceptors.response.use(
-    (response) => response,
-    (error) => {
-      console.error(error);
-      if (error.response?.status === 401) {
-        localStorage.clear();
-        Swal.fire({
-          title: "Sesión expirada",
-          text: "Tu sesión ha vencido. Por favor, inicia sesión nuevamente.",
-          icon: "warning",
-          confirmButtonText: "Aceptar",
-        }).then(() => {
-          router.push("/authentication/login");
-        });
+  if (window && typeof window !== "undefined") {
+    axiosClient.interceptors.response.use(
+      (response) => response,
+      (error) => {
+        console.error(error);
+        if (error.response?.status === 401) {
+          localStorage.clear();
+          Swal.fire({
+            title: "Sesión expirada",
+            text: "Tu sesión ha vencido. Por favor, inicia sesión nuevamente.",
+            icon: "warning",
+            confirmButtonText: "Aceptar",
+          }).then(() => {
+            router.push("/authentication/login");
+          });
+        }
+        return Promise.reject(error);
       }
-      return Promise.reject(error);
-    }
-  );
+    );
+  }
 };
 
 // instancia separada sin interceptores
